refactor(processor): extract widget caption loader
Move lazy widget appearance caption extraction into a focused processor helper. This removes direct file-reading and widget-parser concerns from processDocument while preserving the cached caption lookup behavior.
